Fix #248 by putting `#include "config.h"` in most src/*.c files. (#249)
* Fix #248 by putting `#include "config.h"` in all src/*.c files.

The iwyu tool doesn't handle `config.h` files well, and these includes were
incorrectly removed, which breaks things on some platforms. Add them back to
most `src/*.c` files with `/* IWYU pragma: keep */` to make iwyu ignore them.
We skip `src/hashtable.c` because it is a standalone tool that is platform
independent.

Also add `/* IWYU pragma: keep */` to includes in `src/fileutil.c` that are
needed on some platforms but not others. This means we can remove the special
exemptions to skip this file for the `iwyu` and `iwyu-fix` targets in
`CMakeLists.txt`.

Add some explicit typecasts to `rollsum.[ch]`  and `patch.c` to silence warnings on windows.

Update NEWS.md for `config.h` include fixes and added typecasts.

+++ src/rollsum.h
@@ -52,7 +52,7 @@ static inline void RollsumRotate(Rollsum *sum, unsigned char out,
                                  unsigned char in)
 {
     sum->s1 += in - out;
-    sum->s2 += sum->s1 - sum->count * (out + ROLLSUM_CHAR_OFFSET);
+    sum->s2 += sum->s1 - (uint_fast16_t)sum->count * (out + ROLLSUM_CHAR_OFFSET);
 }
 
 static inline void RollsumRollin(Rollsum *sum, unsigned char in)
@@ -65,7 +65,7 @@ static inline void RollsumRollin(Rollsum *sum, unsigned char in)
 static inline void RollsumRollout(Rollsum *sum, unsigned char out)
 {
     sum->s1 -= out + ROLLSUM_CHAR_OFFSET;
-    sum->s2 -= sum->count * (out + ROLLSUM_CHAR_OFFSET);
+    sum->s2 -= (uint_fast16_t)sum->count * (out + ROLLSUM_CHAR_OFFSET);
     sum->count--;
 }
